requirements.txt
python-dotenv
Flask
flask-smorest
flask-pymongo
pymongo
pandas
marshmallow